requirements.txt
fastapi[all]
requests
uvicorn[standard]
fastapi-utils
psutil
prometheus-fastapi-instrumentator
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-otlp
